Rename ui to presentation
