+++ app/src/main/java/com/alongo/manyweather/di/ActivityBuilderModule.kt
@@ -1,6 +1,6 @@
 package com.alongo.manyweather.di
 
-import com.alongo.manyweather.ui.main.MainActivity
+import com.alongo.manyweather.presentation.main.MainActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
